+++ projects/frontend/es-app/src/guis/menus/GuiMenuVirtualSystems.cpp
@@ -21,6 +21,7 @@ GuiMenuVirtualSystems::GuiMenuVirtualSystems(WindowManager& window, SystemManage
   , mMultiplayersOriginalValues(RecalboxConf::Instance().GetCollectionMultiplayer())
   , mLastPlayedOriginalValues(RecalboxConf::Instance().GetCollectionLastPlayed())
   , mLightGunOriginalValues(!RecalboxConf::Instance().GetCollectionHide("lightgun"))
+  , mPortsOriginalValues(!RecalboxConf::Instance().GetCollectionHide("ports"))
 {
   // All games
   mAllGames = AddSwitch(_("SHOW ALL-GAMES SYSTEM"), mAllGamesOriginalValues, (int)Components::AllGames, this, _(MENUMESSAGE_ADVANCED_ALLGAMES_HELP_MSG));
@@ -31,10 +32,13 @@ GuiMenuVirtualSystems::GuiMenuVirtualSystems(WindowManager& window, SystemManage
   // Last Played
   mLastPlayed = AddSwitch(_("SHOW LAST-PLAYED SYSTEM"), mLastPlayedOriginalValues, (int)Components::LastPlayed, this, _(MENUMESSAGE_ADVANCED_LASTPLAYED_HELP_MSG));
 
-  // Last Played
+  // Lightgun
   BoardType board = Board::Instance().GetBoardType();
   if (board != BoardType::OdroidAdvanceGo && board != BoardType::OdroidAdvanceGoSuper)
-    mLastPlayed = AddSwitch(_("SHOW LIGHTGUN SYSTEM"), mLightGunOriginalValues, (int)Components::LightGun, this, _(MENUMESSAGE_ADVANCED_LIGHTGUN_HELP_MSG));
+    mLightGun = AddSwitch(_("SHOW LIGHTGUN SYSTEM"), mLightGunOriginalValues, (int)Components::LightGun, this, _(MENUMESSAGE_ADVANCED_LIGHTGUN_HELP_MSG));
+
+  // Ports
+  mPorts = AddSwitch(_("SHOW PORTS SYSTEM"), mPortsOriginalValues, (int)Components::Ports, this, _(MENUMESSAGE_ADVANCED_PORTS_HELP_MSG));
 
   AddSubMenu(_("VIRTUAL SYSTEMS PER GENRE"), (int)Components::VirtualPerGenre, _(MENUMESSAGE_ADVANCED_VIRTUALGENRESYSTEMS_HELP_MSG));
   AddSubMenu(_("ARCADE VIRTUAL SYSTEM"), (int)Components::VirtualArcade, _(MENUMESSAGE_ADVANCED_ARCADEVIRTUALSYSTEM_HELP_MSG));
@@ -45,6 +49,7 @@ GuiMenuVirtualSystems::~GuiMenuVirtualSystems()
   if ((mLastPlayedOriginalValues != RecalboxConf::Instance().GetCollectionLastPlayed()) ||
       (mMultiplayersOriginalValues != RecalboxConf::Instance().GetCollectionMultiplayer()) ||
       (mLightGunOriginalValues == RecalboxConf::Instance().GetCollectionHide("lightgun")) ||
+      (mPortsOriginalValues == RecalboxConf::Instance().GetCollectionHide("ports")) ||
       (mAllGamesOriginalValues != RecalboxConf::Instance().GetCollectionAllGames()))
     RequestRelaunch();
 }
@@ -63,6 +68,7 @@ void GuiMenuVirtualSystems::SwitchComponentChanged(int id, bool status)
     case Components::Multiplayers: RecalboxConf::Instance().SetCollectionMultiplayer(status).Save(); break;
     case Components::LastPlayed: RecalboxConf::Instance().SetCollectionLastPlayed(status).Save(); break;
     case Components::LightGun: RecalboxConf::Instance().SetCollectionHide("lightgun", !status).Save(); break;
+    case Components::Ports: RecalboxConf::Instance().SetCollectionHide("ports", !status).Save(); break;
     case Components::VirtualPerGenre:
     case Components::VirtualArcade: break;
   }